daemon: newDaemon returns promise for *option.DaemonConfig.
This resolves in the same way that the daemon promise resolves, but
resolves to *option.DaemonConfig notifying that newDaemon init tasks
have completed. Providing daemon config as a resolvable promise
allows code that is being added to the Daemon cells to wait for this
resolution without introducing cyclic package dependencies.

This is necessary as newDaemon may override fields in DaemonConfig such
as when doing kube proxy replacement init or device detection.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>