metrics/features: introduce features package to track enable features
[ upstream commit 7a5852f3cf31bd3cfc9d81991f1fc11d6126f8d1 ]

With a dedicate package to detect which features are enabled, we will be
able to report, in form of metrics, from the agents, which features are
enabled. Once all features are tracked, it will give us the ability to
track down which features are being tested in our CI and the ones that
are lacking coverage.

Signed-off-by: André Martins <andre@cilium.io>