ipsec: rework encryption connectivity test for VinE traffic
IPsec VinE traffic can follow the same testing patterns as Wireguard
sans the Host-to-Host encryption parts.

Update the encryption connectivity to perform a more abstract "Encrypted
Overlay" check. This is an general term which expresses that traffic is first
tunnel encapsulated and then it is encrypted. With IPsec's move to VinE
traffic this means both WG and IPsec support encrypted overlay past
specific versions of Cilium. If the check is true then tests will sniff
on the native device and expect to see no plain tunnel packets.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>